Move MapGet extensions to own file
